+++ pomp/core/item.py
@@ -22,12 +22,22 @@ class Item(OrderedDict):
         for key in sorted(fields_by_creation_counter):
             field = fields_by_creation_counter[key]
             value = _args.pop() if _args else kwargs.get(field, None)
+            print("set", key, field, value)
             super(Item, self).__setitem__(field, value)
 
     def __setattr__(self, key, value):
         if key in self:
             super(Item, self).__setitem__(key, value)
         super(Item, self).__setattr__(key, value)
+
+    def __getattribute__(self, key):
+        value = super(Item, self).__getattribute__(key)
+
+        # if attribute is a item field return it`s value, not a Field object
+        if isinstance(value, Field):
+            return self[key]
+
+        return value
 
 
 class Field(object):